refactor: decompose studio_handlers.py into _studio/ subpackage (Phase 1)
Extract 53 helper functions from the 2,717 LOC studio_handlers.py into 11
focused modules under routes/_studio/:

- ui_utils.py: toast helpers, type coercion
- manifest_helpers.py: manifest loading, source resolution, canvas helpers
- context.py: studio context normalization, recent hub, tab management
- thumbnail_cache.py: thumbnail sizing, pruning, remote cache management
- page_job_prefs.py: highres/stitch job preference persistence
- page_source_prefs.py: page source preference resolution
- download_feedback.py: download/highres/stitch/optimize feedback
- scan_resolution.py: local scan info, remote dimension probing
- export_thumbnails.py: export thumbnail building and slice assembly
- workspace.py: workspace context + tab content building

studio_handlers.py is now 1,045 LOC containing only route handlers.

Part of #114
